delete remote video file upon complete file transfer (total byte size)

+++ src/devices/adb/index.js
@@ -41,20 +41,23 @@ const startCameraApp = (device) => {
 
 export const pullVideoFile = (data) => {
     const device = data.device
+    const remoteFileSize = data.fileSize
     const pullFilePath = data.pullFilePath
     const timestamp = data.timestamp
     const endTime = data.endTime
     const videoFileName = path.basename(pullFilePath)
     const mediaDir = global.podware.currentRecordingSession.mediaDir
     const outFile = mediaDir + videoFileName
-        spawn('adb',['-s',device,'pull',pullFilePath,outFile]).stdout.on('data',data => {
-            /*
-            if(fs.existsSync(outFile)) {
+    spawn('adb',['-s',device,'pull',pullFilePath,outFile]).stdout.on('data',data => {
+        var flag = true
+        while(flag) {
+            if(fs.existsSync(outFile) && fs.statSync(outFile).size == remoteFileSize) {
                 console.log(videoFileName + ' exists locally -> now deleting from ' + device)
                 exec('adb -s ' + device + ' shell rm -rf ' + pullFilePath, (err,stdout,stdin) => {
                     if(err) console.log(err)
                 })
+                flag = false
             }
-            */
+        }
     })
 }
